Start using astc codec

+++ src/libs/astc/astc_compress_symbolic.cpp
@@ -833,8 +833,8 @@ void expand_block_artifact_suppression(int xdim, int ydim, int zdim, error_weigh
 				float zdif = (z - centerpos_z) / zdim;
 
 				float wdif = 0.36f;
-				float dist = sqrt(xdif * xdif + ydif * ydif + zdif * zdif + wdif * wdif);
-				*bef = pow(dist, ewp->block_artifact_suppression);
+				float dist = (float)sqrt(xdif * xdif + ydif * ydif + zdif * zdif + wdif * wdif);
+				*bef = (float)pow(dist, ewp->block_artifact_suppression);
 				bef++;
 			}
 }
@@ -907,10 +907,10 @@ float prepare_error_weight_block(const astc_codec_image * input_image,
 						avg.xyz = float3(favg, favg, favg) * mixing + avg.xyz * (1.0f - mixing);
 						variance.xyz = float3(fvar, fvar, fvar) * mixing + variance.xyz * (1.0f - mixing);
 
-						float4 stdev = float4(sqrt(MAX(variance.x, 0.0f)),
-											  sqrt(MAX(variance.y, 0.0f)),
-											  sqrt(MAX(variance.z, 0.0f)),
-											  sqrt(MAX(variance.w, 0.0f)));
+						float4 stdev = float4((float)sqrt(MAX(variance.x, 0.0f)),
+                                              (float)sqrt(MAX(variance.y, 0.0f)),
+                                              (float)sqrt(MAX(variance.z, 0.0f)),
+                                              (float)sqrt(MAX(variance.w, 0.0f)));
 
 						avg.xyz = avg.xyz * ewp->rgb_mean_weight;
 						avg.w = avg.w * ewp->alpha_mean_weight;
@@ -958,15 +958,15 @@ float prepare_error_weight_block(const astc_codec_image * input_image,
 						if (r < 0.0031308f)
 							r = 12.92f;
 						else
-							r = 0.4396f * pow(r, -0.58333f);
+							r = 0.4396f * float(pow(r, -0.58333f));
 						if (g < 0.0031308f)
 							g = 12.92f;
 						else
-							g = 0.4396f * pow(g, -0.58333f);
+							g = 0.4396f * float(pow(g, -0.58333f));
 						if (b < 0.0031308f)
 							b = 12.92f;
 						else
-							b = 0.4396f * pow(b, -0.58333f);
+							b = 0.4396f * float(pow(b, -0.58333f));
 						error_weight.x *= r;
 						error_weight.y *= g;
 						error_weight.z *= b;
@@ -1119,12 +1119,12 @@ void prepare_block_statistics(int xdim, int ydim, int zdim, const imageblock * b
 	float bb_var = cov_matrix.v[2].z;
 	float aa_var = cov_matrix.v[3].w;
 
-	float rg_correlation = cov_matrix.v[0].y / sqrt(MAX(rr_var * gg_var, 1e-30f));
-	float rb_correlation = cov_matrix.v[0].z / sqrt(MAX(rr_var * bb_var, 1e-30f));
-	float ra_correlation = cov_matrix.v[0].w / sqrt(MAX(rr_var * aa_var, 1e-30f));
-	float gb_correlation = cov_matrix.v[1].z / sqrt(MAX(gg_var * bb_var, 1e-30f));
-	float ga_correlation = cov_matrix.v[1].w / sqrt(MAX(gg_var * aa_var, 1e-30f));
-	float ba_correlation = cov_matrix.v[2].w / sqrt(MAX(bb_var * aa_var, 1e-30f));
+	float rg_correlation = cov_matrix.v[0].y / float(sqrt(MAX(rr_var * gg_var, 1e-30f)));
+	float rb_correlation = cov_matrix.v[0].z / float(sqrt(MAX(rr_var * bb_var, 1e-30f)));
+	float ra_correlation = cov_matrix.v[0].w / float(sqrt(MAX(rr_var * aa_var, 1e-30f)));
+	float gb_correlation = cov_matrix.v[1].z / float(sqrt(MAX(gg_var * bb_var, 1e-30f)));
+	float ga_correlation = cov_matrix.v[1].w / float(sqrt(MAX(gg_var * aa_var, 1e-30f)));
+	float ba_correlation = cov_matrix.v[2].w / float(sqrt(MAX(bb_var * aa_var, 1e-30f)));
 
 	if (astc_isnan(rg_correlation))
 		rg_correlation = 1.0f;
@@ -1139,11 +1139,11 @@ void prepare_block_statistics(int xdim, int ydim, int zdim, const imageblock * b
 	if (astc_isnan(ba_correlation))
 		ba_correlation = 1.0f;
 
-	float lowest_correlation = MIN(fabs(rg_correlation), fabs(rb_correlation));
-	lowest_correlation = MIN(lowest_correlation, fabs(ra_correlation));
-	lowest_correlation = MIN(lowest_correlation, fabs(gb_correlation));
-	lowest_correlation = MIN(lowest_correlation, fabs(ga_correlation));
-	lowest_correlation = MIN(lowest_correlation, fabs(ba_correlation));
+	float lowest_correlation = (float)MIN(fabs(rg_correlation), fabs(rb_correlation));
+	lowest_correlation = (float)MIN(lowest_correlation, fabs(ra_correlation));
+	lowest_correlation = (float)MIN(lowest_correlation, fabs(gb_correlation));
+	lowest_correlation = (float)MIN(lowest_correlation, fabs(ga_correlation));
+	lowest_correlation = (float)MIN(lowest_correlation, fabs(ba_correlation));
 	*lowest_correl = lowest_correlation;
 
 	// compute a "normal-map" factor
@@ -1162,7 +1162,7 @@ void prepare_block_statistics(int xdim, int ydim, int zdim, const imageblock * b
 							blk->orig_data[4 * i + 2]);
 		val = (val - float3(0.5f, 0.5f, 0.5f)) * 2.0f;
 		float length_squared = dot(val, val);
-		float nf = fabs(length_squared - 1.0f);
+		float nf = (float)fabs(length_squared - 1.0f);
 		nf_sum += nf;
 	}
 	float nf_avg = nf_sum / texels_per_block;
